server/requirements.txt
Django==5.1.4
google-generativeai==0.8.3
requests==2.31.0